post

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,25 @@ class TasksController extends Controller
      */
     public function index()
     {
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderby('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         //タスクの一覧を取得
-        $tasks = Task::orderBy('id', 'desc')->paginate(25);
+        // $tasks = Task::orderBy('id', 'desc')->paginate(25);
         
         //タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        // return view('tasks.index', [
+        //     'tasks' => $tasks,
+        //     ]);
+        
+        return view('welcome', $data);
     }
 
     /**
@@ -53,11 +65,15 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            
         //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $task = new Task,
+        $task->status = $request->status,
+        $task->content = $request->content,
+        $task->save(),
+            ]);
         
         //トップページへリダイレクトさせる
         return redirect('/');
@@ -133,10 +149,15 @@ class TasksController extends Controller
     {
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            
         //メッセージを削除
         $task->delete();
+        }
         
         //トップページへリダイレクトさせる
         return redirect('/');
+        }
     }
-}
